primer acercamiento a una solucion concisa
hacen falta solucionar varios errores. entre ellos el porque no me deja implementar correctamente los @Bean

+++ src/main/java/com/BootcampPragma/Api_Emazon/application/dto/ArticuloDto.java
@@ -14,6 +14,6 @@ public class ArticuloDto {
     private String nombre;
     private String descripcion;
     private long cantidad;
-    private long precio;
+    private double precio;
 
 }
